+++ oggm/core/climate.py
@@ -977,7 +977,28 @@ def calving_mb(gdir):
     return gdir.inversion_calving_rate * 1e9 * rho / gdir.rgi_area_m2
 
 
-@entity_task(log, writes=['local_mustar'])
+def _fallback_local_t_star(gdir):
+    """A Fallback function if climate.local_t_star raises an Error.
+
+    This function will still write a `local_mustar.json`, filled with NANs,
+    if climate.local_t_star fails and cfg.PARAMS['continue_on_error'] = True.
+
+    Parameters
+    ----------
+    gdir : :py:class:`oggm.GlacierDirectory`
+        the glacier directory to process
+
+    """
+    # Scalars in a small dict for later
+    df = dict()
+    df['rgi_id'] = gdir.rgi_id
+    df['t_star'] = np.nan
+    df['bias'] = np.nan
+    df['mu_star_glacierwide'] = np.nan
+    gdir.write_json(df, 'local_mustar')
+
+
+@entity_task(log, writes=['local_mustar'], fallback=_fallback_local_t_star)
 def local_t_star(gdir, *, ref_df=None, tstar=None, bias=None):
     """Compute the local t* and associated glacier-wide mu*.